test(client-preset): split skipIndexFile tests into their own spec file
Per review feedback on PR #10946: move the skipIndexFile tests out of
client-preset.spec.ts into client-preset.presetConfig.skipIndexFile.spec.ts,
matching the existing per-option spec split (e.g.
client-preset.fragment-masking.spec.ts), with the describe block renamed
to 'client-preset - presetConfig.skipIndexFile'.
